requirements.txt
Flask
Flask-Login
Flask-WTF
Werkzeug
mysql-connector-python
nltk
reportlab
python-dotenv
